feat: add tests for all CLI commands shown in docs
This makes use of the newly-published
[soroban-test](https://docs.rs/soroban-test/latest/soroban_test/) crate
and the newly-reorganized [soroban-cli](https://docs.rs/soroban-cli)
library to test all commands shown in the docs.

This gives us confidence that all of these commands work, though it does
come with some downsides:

- these tests live far from the thing they are meant to test
  - could we generate the docs from this repository, rather than keeping
    them in sync with each other?
  - maybe, but it's a little unclear how that would work
  - this is a good starting point

Another shortcoming of the approach here: it still makes use of
`assert_cmd` in some tests, which runs the version of `soroban-cli`
installed on the host system, making it quite brittle. This is necessary
because some commands (`Cmd`s) from `soroban-cli` do not currently
operate in a test-friendly way. This can be seen in a secondary way,
too: not all `Cmd`s can be run the same way. Some use `run`, some use
`run_in_sandbox`, some take arguments, some do not. These concerns can
all be addressed together by refactoring `soroban-cli` to make all
`Cmd`s implement an asynchronous `run` method with the same signature,
and which will return a struct containing both `stdout` and `stderr`.

However, even with these shortcomings and downsides, it is still worth
merging these tests as-is, since it will give us higher confidence that
everything works as expected with the current code. We can refactor
these tests later, and potentially move them to a different repo.

Co-authored-by: Chad Ostrowski <221614+chadoh@users.noreply.github.com>