Ensure deposited token amounts both > 0
For the initial deposit, if one of the deposited token amounts is 0, the
contract can get into a situationw here one of the reserves is 0,
whereas the other is positive. This prevents all further deposits.
See the previous commit for a test case.

The contract is written to assume that `reserve_a > 0 && reserve_b > 0`
after the initial deposit (unless all shares have been withdrawn). For
instance, this assumption shown up in `deposit` when `new_total_shares`
is computed, and also in `get_deposit_amounts`.

This commit actually enforces this invariant.