Enforce reserve amounts amounts strictly positive (#290)
* Add test to trigger divide by zero in LP contract

* Ensure deposited token amounts both > 0

For the initial deposit, if one of the deposited token amounts is 0, the
contract can get into a situationw here one of the reserves is 0,
whereas the other is positive. This prevents all further deposits.
See the previous commit for a test case.

The contract is written to assume that `reserve_a > 0 && reserve_b > 0`
after the initial deposit (unless all shares have been withdrawn). For
instance, this assumption shown up in `deposit` when `new_total_shares`
is computed, and also in `get_deposit_amounts`.

This commit actually enforces this invariant.

* Add test to prevent all deposits using a swap

The same situation in which one reserve is 0 whereas the other is not
(and is positive) can be reached using a `swap`, as this test case
shows. This prevents all further deposits.

* Enforce in `swap` that new reserves are positive

Without this, it is possible through a `swap` to prevent all further
deposits. See the previous commit for a test case demonstrating this.

---------

Co-authored-by: Siddharth Suresh <siddharth@stellar.org>